@@ -1,15 +1,13 @@
 {
-  "name": "@deia/vue-visjs",
-  "version": "0.6.4",
+  "name": "vue3-visjs",
+  "version": "0.7.0",
   "private": false,
   "main": "./dist/vue-visjs.ssr.js",
   "module": "./dist/vue-visjs.esm.js",
   "unpkg": "./dist/vue-visjs.min.js",
   "scripts": {
-    "serve": "vue-cli-service serve",
     "build": "rollup --config build/rollup.config.js --environment BUILD:production",
-    "test:unit": "vue-cli-service test:unit",
-    "lint": "vue-cli-service lint",
+    "build:es": "rollup --config build/rollup.config.js --format es --environment BUILD:production --file dist/vue-visjs.esm.js",
     "beforePublish": "npm run lint && npm run build",
     "prepare": "npm run build"
   },
@@ -19,12 +17,13 @@
   ],
   "repository": {
     "type": "git",
-    "url": "git://github.com/sjmallon/vue-visjs.git"
+    "url": "https://github.com/alexdeia/vue3-visjs"
   },
-  "author": "Steve Mallon <public.sjmallon@gmail.com>",
+  "author": "Aleksei Klykov <alekx.deia@gmail.com>",
   "license": "MIT",
   "keywords": [
     "vue",
+    "vue3",
     "vis",
     "visjs"
   ],
@@ -37,15 +36,15 @@
     "moment": "^2.27.0",
     "propagating-hammerjs": "^1.4.7",
     "uuid": "^8.3.2",
-    "vis-data": "^7.1.4",
+    "vis-data": "7.1.3",
     "vis-network": "^9.1.2",
     "vis-util": "^5.0.3",
-    "vue": "~2.x",
+    "vue": "^3.2.45",
     "vue-runtime-helpers": "^1.1.2",
     "xss": "^1.0.8"
   },
   "peerDependencies": {
-    "vue": "~2.x"
+    "vue": "^3.2.45"
   },
   "devDependencies": {
     "@babel/core": "^7.14.0",
@@ -56,27 +55,28 @@
     "@rollup/plugin-commonjs": "^18.1.0",
     "@rollup/plugin-node-resolve": "^13.0.0",
     "@rollup/plugin-replace": "^2.4.2",
-    "@vue/cli-plugin-babel": "^4.4.6",
-    "@vue/cli-plugin-eslint": "^4.5.12",
-    "@vue/cli-plugin-unit-jest": "^4.5.12",
-    "@vue/cli-service": "^4.1.1",
-    "@vue/eslint-config-prettier": "^5.0.0",
-    "@vue/test-utils": "^1.0.3",
+    "@vue/compiler-sfc": "^3.2.45",
+    "@vue/eslint-config-prettier": "^7.0.0",
+    "@vue/test-utils": "^2.2.8",
+    "autoprefixer": "^10.4.13",
     "babel-eslint": "^10.1.0",
     "canvas": "^2.7.0",
+    "component-emitter": "^1.3.0",
     "eslint": "^6.8.0",
     "eslint-plugin-prettier": "^3.1.4",
     "eslint-plugin-vue": "^7.9.0",
     "fibers": "^4.0.3",
     "minimist": "^1.2.5",
-    "node-webcrypto-ossl": "^2.1.0",
     "postcss": "^8.2.13",
-    "prettier": "^1.18.2",
+    "prettier": "^2.8.3",
     "rollup": "^2.18.2",
-    "rollup-plugin-postcss": "^3.1.8",
+    "rollup-plugin-postcss": "^4.0.2",
+    "rollup-plugin-scss": "3",
     "rollup-plugin-terser": "^7.0.2",
-    "rollup-plugin-vue": "^5.1.9",
-    "sass": "^1.26.9",
-    "sass-loader": "^8.0.2"
+    "rollup-plugin-vue": "^6.0.0",
+    "sass": "^1.57.1",
+    "sass-loader": "^8.0.2",
+    "timsort": "^0.3.0",
+    "vue-loader": "^17.0.1"
   }
 }
