@@ -5,29 +5,27 @@ import '../node_modules/@deia/vis-timeline/styles/vis-timeline-graph2d.min.css'
 import '../node_modules/vis-network/styles/vis-network.min.css'
 
 // Declare install function executed by Vue.use()
-export function install(Vue) {
+export function install(app) {
   if (install.installed) return
+
   install.installed = true
-  Vue.component(Timeline.name, Timeline)
-  Vue.component(Network.name, Network)
-  Vue.component(Graph2D.name, Graph2d)
+  app.component(Timeline.name, Timeline)
+  app.component(Network.name, Network)
+  app.component(Graph2d.name, Graph2d)
 }
 
-// Create module definition for Vue.use()
-const plugin = {
-  install
-}
+const plugin = { install }
 
-// Auto-install when vue is found (eg. in browser via <script> tag)
+// To auto-install when Vue is found
 let GlobalVue = null
 if (typeof window !== 'undefined') {
   GlobalVue = window.Vue
 } else if (typeof global !== 'undefined') {
-  /* eslint no-undef: 0 */
   GlobalVue = global.Vue
 }
 if (GlobalVue) {
   GlobalVue.use(plugin)
 }
+
 export { DataSet, DataView } from 'vis-data/esnext'
 export { Timeline, Graph2d, Network }
